requirements.txt
uvicorn
fastapi
pillow
numpy
pandas
deepface
openpyxl
python-multipart
